+++ src/main/java/com/fxsong/demo/crud/member/MemberRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.ListCrudRepository;
 
 import java.util.List;
 
-public interface MemberRepository extends ListCrudRepository<Member, Integer> {
+public interface MemberRepository extends ListCrudRepository<Member, Long> {
     List<Member> findByName(String name);
 }
